Pozmieniane błędy w konstrukcji projektu, aby był bardziej czytelny typu usunięte walidacje w repozytoriach przeniesione wszystkie walidacje bez regexow i permisji do serwisów AuctionsController w wersji działającej prototypowej

+++ api/Services/AuctionS/AuctionService.cs
@@ -1,35 +1,61 @@
-﻿using api.Models;
+﻿//Miejsce w którym program waliduje dane, koordynuje działanie respozytorium i transformuje dane np mapuje dto na model
+
+using api.Models;
+using api.DTOs.AuctionDto;
 using api.Repositories.AuctionRepo;
-using NuGet.Protocol.Core.Types;
 
 namespace api.Services.AuctionS
 {
-    public class AuctionService : IAuctionService
+    public class AuctionService(IAuctionRepository repository) : IAuctionService
     {
-        private readonly IAuctionRepository _repository;
+        private readonly IAuctionRepository _repository = repository;
 
-        public AuctionService(IAuctionRepository repository)
-        {
-            _repository = repository;
-        }
         public async Task<IEnumerable<Auction>> GetAllAsync()
         {
             return await _repository.GetAllAsync();
         }
         public async Task<Auction?> GetByIdAsync(long id)
         {
-            return await _repository.GetByIdAsync(id);
+            var auction = await _repository.GetByIdAsync(id);
+            return auction ?? throw new KeyNotFoundException($"Auction with ID: {id} not found");
         }
-        public async Task<Auction> AddAuctionAsync(Auction auction)
+        public async Task<Auction> AddAuctionAsync(AuctionCreateDto dto)
         {
-            return await _repository.AddAuctionAsync(auction);
+            var newAuction = new Auction
+            {
+                Name = dto.Name,
+                Description = dto.Description,
+                Category = dto.Category,
+                Price = dto.Price,
+                EndAt = dto.EndAt
+            };
+            if (newAuction.EndAt < DateTime.UtcNow)
+            {
+                throw new InvalidOperationException("End date must be in the future");
+            }
+            return await _repository.AddAuctionAsync(newAuction);
         }
-        public async Task<Auction> UpdateAuctionAsync(Auction auction)
+        public async Task<Auction> UpdateAuctionAsync(long id, AuctionUpdateDto dto)
         {
+            var auction = await _repository.GetByIdAsync(id);
+            if (auction!.Bids.Count != 0)
+            {
+                throw new InvalidOperationException("Cannot change price when bids already exist");
+            }
+            auction.Name = dto.Name ?? auction.Name;
+            auction.Description = dto.Description ?? auction.Description;
+            auction.Category = dto.Category ?? auction.Category;
+            auction.Price = dto.Price ?? auction.Price;
+
             return await _repository.UpdateAuctionAsync(auction);
         }
         public async Task DeleteAuctionAsync(long id)
         {
+            var auction = await _repository.GetByIdAsync(id);
+            if(auction!.Bids.Count != 0)
+            {
+                throw new InvalidCastException("Cannot delete Auction when bids already exist");
+            }
             await _repository.DeleteAuctionAsync(id);
         }
     }
